chore: move OpenCV Mat `use` util package for readability
